+++ PoltergeistEngine/include/PoltergeistEngine/Rendering/Texture.hpp
@@ -14,7 +14,7 @@ private:
 public:
 	~Texture() noexcept;
 
-	[[nodiscard]] static std::shared_ptr<Texture> CreateEmpty(size_t width, size_t height, uint8_t slot);
+	[[nodiscard]] static std::shared_ptr<Texture> CreateEmpty(uint32_t width, uint32_t height, uint8_t slot);
 	[[nodiscard]] static std::shared_ptr<Texture> CreateFromFile(const std::filesystem::path& texturePath, uint8_t slot);
 
 	void Bind() const noexcept;
